Strict GNU C23 toolchain: compiler contract, warning tiers, and analyzer fixes (#185)
Two-compiler contract (GCC 13/Clang 18 minimum, GCC 16.2/Clang 22.1.8 current) with a strict -Werror matrix, one probed warning-tier flag list shared by Autotools and CMake, the migration warning budget burned down to zero, a scheduled GCC analyzer workflow, and fixes for the defects the analyzer and the stricter warnings exposed.

Refs #86.

+++ src/act.comm.do_spec_comm.c
@@ -63,7 +63,7 @@ ACMD(do_spec_comm)
 
     if (subcmd == SCMD_ASK)
     {
-      len = strlen(buf2);
+      len = (int)strlen(buf2);
       // remove trailing punctuation from ask
       while (len >= 0 && (buf2[len - 1] == '.' || buf2[len - 1] == '!' || *buf2 == '\n'))
       {
