+++ src/random.c
@@ -51,8 +51,8 @@ unsigned long circle_random(void)
 {
   int lo, hi, test;
 
-  hi = seed / q;
-  lo = seed % q;
+  hi = (int)(seed / q);
+  lo = (int)(seed % q);
 
   test = a * lo - r * hi;
