+++ src/wilderness/resource_depletion.h
@@ -15,15 +15,15 @@
 void init_resource_depletion_database(void);
 
 /* Basic depletion functions */
-float get_resource_depletion_rate(int resource_type);
-float get_resource_depletion_level(room_rnum room, int resource_type);
-float get_resource_depletion_level_by_coords(int x, int y, int zone_vnum, int resource_type);
+double get_resource_depletion_rate(int resource_type);
+double get_resource_depletion_level(room_rnum room, int resource_type);
+double get_resource_depletion_level_by_coords(int x, int y, int zone_vnum, int resource_type);
 void apply_harvest_depletion(room_rnum room, int resource_type, int quantity);
 
 /* Phase 7: Enhanced depletion with cascade effects */
 void apply_harvest_depletion_with_cascades(room_rnum room, int resource_type, int quantity);
 void apply_cascade_effects(room_rnum room, int source_resource, int quantity);
-void apply_single_cascade_effect(room_rnum room, int target_resource, float effect_magnitude,
+void apply_single_cascade_effect(room_rnum room, int target_resource, double effect_magnitude,
                                  const char *description);
 
 /* Phase 7: Cascade preview and analysis */
@@ -34,18 +34,18 @@ int get_ecosystem_state(room_rnum room);
 void show_ecosystem_analysis(struct char_data *ch, room_rnum room);
 
 /* Regeneration functions */
-float get_resource_regeneration_rate(int resource_type);
-float get_modified_regeneration_rate(int resource_type, int x, int y);
-float calculate_regeneration_amount(int resource_type, time_t last_harvest_time, int x, int y);
+double get_resource_regeneration_rate(int resource_type);
+double get_modified_regeneration_rate(int resource_type, int x, int y);
+double calculate_regeneration_amount(int resource_type, time_t last_harvest_time, int x, int y);
 void apply_lazy_regeneration(room_rnum room, int resource_type);
 bool should_harvest_fail_due_to_depletion(room_rnum room, int resource_type);
-float get_harvest_success_modifier(room_rnum room, int resource_type);
-const char *get_depletion_level_name(float resource_level);
+double get_harvest_success_modifier(room_rnum room, int resource_type);
+const char *get_depletion_level_name(double resource_level);
 
 /* Conservation functions (stubs for now) */
 void update_conservation_score(struct char_data *ch, int resource_type, bool sustainable);
-float get_player_conservation_score(struct char_data *ch);
-const char *get_conservation_status_name(float score);
+double get_player_conservation_score(struct char_data *ch);
+const char *get_conservation_status_name(double score);
 void show_resource_conservation_status(struct char_data *ch, int x, int y);
 void show_regeneration_analysis(struct char_data *ch, int x, int y);
